Started adapting Python to Corba
The Python events only supported numeric values, and in Python measurements
and events are completely separate (duplicate). Started adding events of
different types and merging events and measurements as they are in corba.
This is not finished: need to rewrite event() and meas() in timing.cpp and
make measurements() in Parser a convenience function.

Also, fixed more smallish problems with CORBA not compiling under -Wall -Werror.
--
Olaf


git-svn-id: https://atom-svn.stanford.edu/svn/FPGA/sti@120 8bcac300-4d60-4aec-84ba-b7489f70e69b