Cleaned up Python measurements.
This changes how the Parser class presents measurements to the user-code.
Please be aware that you will have to minimally adopt any code that already
used Parser::measurements(). That aside, I think that this is much cleaner
and nicer.

The data-structures now fully support events of string or DDS-typle type,
but this is not yet implemented in the code for the Python command event().

Also, I accidentially recreated the CORBA-files. Some ID strings seem to have
changed, but nothing else. If this happens more often, I will find a way to
prevent unneeded changes / commits. For now, please ignore the differences.
(^_^;)

--
Olaf


git-svn-id: https://atom-svn.stanford.edu/svn/FPGA/sti@125 8bcac300-4d60-4aec-84ba-b7489f70e69b