STI_Device library and testDevice compiles again in Linux, with a hack
* Changed device/src/GNUmakefile.am to point to the correct include 
files in device/include

* WARNING: There's still a problem with automake-idl. Right now 
automake-idl is using the wrong extensions for idl-generated files.  It 
should make device.h, but instead it insists on making making device.hpp 
and deviceServer.hpp.  All the source code has #include "device.h", and 
this file lives in device/include since it's on the svn server.  This 
means that the device.* code generated by automake-idl is not used by 
the rest of the code, so changing the idl will break the code.

* The output binary files still do not end up in the right directory.  
I want to have the libcorba.a library end up in device/lib, but it 
always gets put in device/src.  This is also true for testDevice.  There 
must be a way to tell g++ the location of the output.

Jason Hogan




git-svn-id: https://atom-svn.stanford.edu/svn/FPGA/sti-main/branches/sti@237 8bcac300-4d60-4aec-84ba-b7489f70e69b