Multiplexer device.
Partner devices and channels (for both the digital control lines and the line to be muxed) are provided through the initialization file. 

Note that all measurements saved will be duplicate a measurement saved by the muxed line, but at least the data is organized according to mux channel for easy reading with GetDeviceMeasurements.

Compiles, mostly runtime-debugged. Most issues seem to stem from using the Analog In.

-Susannah

git-svn-id: https://atom-svn.stanford.edu/svn/FPGA/sti/trunk@1084 8bcac300-4d60-4aec-84ba-b7489f70e69b