First stab at Point Grey camera support. Tested on Windows 7 and 8 (device compiled x64).
* Very limited functionality at this point, but good enough for snapping selfies
* Works with at least two different Chameleon models, a color and a B&W model
* Specify camera by serial number (sadly, too big to store in the STI Module's unsigned short)

As an aside, the Point Grey SDK is quite nice. Not only is talking to the camera easy (functional STI device in less than a day), but they even include a nice Image class that will easily save images in a variety of formats (I've tested .tif, .png and .pgm). And it even has Linux and OS X support, so this code might even be cross-platform without even trying.

git-svn-id: https://atom-svn.stanford.edu/svn/FPGA/sti/trunk@1095 8bcac300-4d60-4aec-84ba-b7489f70e69b