Long overdue commit. Various cleanups, though still a lot of work to be done.
This may require a different version of the rs232 code, though it will probably only take minimal changes to compile it with the existing rs232 code, at the expense of a few small bits of added functionality.

I also added some sample files in doc/, which will hopefully be useful.

git-svn-id: https://atom-svn.stanford.edu/svn/FPGA/sti/trunk@1144 8bcac300-4d60-4aec-84ba-b7489f70e69b