Long overdue commit. Various cleanups, though still a lot of work to be done.
This may require a different version of the rs232 code, though it will probably only take minimal changes to compile it with the existing rs232 code, at the expense of a few small bits of added functionality.

I also added some sample files in doc/, which will hopefully be useful.