Fixing memory leak in event parsing
This bug causes the etrax timing systems to run out of memory and eventual hang after about 100 shots or so.

Two problems corrected:
(1) Missing virtual destructors in a couple of SynchronousEvent class derivatives, including SynchronousEvent.  Added these. (This didn't fix the memory problem)
(2) Missing destructor for SynchronousEvent!  It calls new for 4 objects (1 omni_mutex and 3 omni_condition variables).  These were not being deleted!! Noticed this in the heap profiler.  This fixed the problem.